\name{QuadraticGarch-package}
\alias{QuadraticGarch-package}
\alias{QuadraticGarch}
\docType{package}
\title{QGARCH (1,1) Estimation}
\description{
 This package performs the MLE of the Quadratic GARCH developed by Sentana (1995) with the GED distribution for innovations.}
\details{
\tabular{ll}{
Package: \tab QuadraticGarch\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-08-14\cr
License: \tab What license is it under?\cr
}

}
\author{
Sebastian Cano-Berlanga <cano.berlanga@gmail.com>

}
